Add `BufferedFirestoreContainer`, move some function into it
