+++ dbux-projects/src/backend/containers/UserEventContainer.js
@@ -1,6 +1,6 @@
 import { newLogger } from '@dbux/common/src/log/logger';
 import { onUserEvent } from '../../userEvents';
-import FirestoreContainer from '../FirestoreContainer';
+import BufferedFirestoreContainer from '../BufferedFirestoreContainer';
 
 /** @typedef {import('../db').Db} Db */
 
@@ -9,7 +9,7 @@ const { log, debug, warn, error: logError } = newLogger('UserEventContainer');
 
 const Verbose = true;
 
-export default class UserEventContainer extends FirestoreContainer {
+export default class UserEventContainer extends BufferedFirestoreContainer {
   buffer = [];
 
   /**
@@ -17,12 +17,8 @@ export default class UserEventContainer extends FirestoreContainer {
    */
   constructor(db) {
     super(db, 'userEvents');
-
-    this.buffer = this.db.backendController.practiceManager.externals.storage.get(this._mementoKeyName) || [];
-    this._previousFlushTime = new Date();
-
-    Verbose && debug(`restore buffer @${this.collectionName}`, this.buffer);
   }
+
 
   async init() {
     super.init();
@@ -31,43 +27,6 @@ export default class UserEventContainer extends FirestoreContainer {
     onUserEvent(this.addEvent);
 
     await this.flush();
-
-    // TODO: start listen on dbux-code/src/userEvents
-    // TODO: start listen on dbux-projects/src/userEvents
-    // TODO: write an event handler for when an event occurs:
-    //    - Store the buffer array to DB (1) if it contains at least 100 events or (2) 5 minutes after receiving the first event. Make sure, that buffer never gets lost (use Memento etc.)
-    //      - return this.addDoc(data);
-
-    // NOTE: The maximum document size in Firestore is 1MiB (see: https://firebase.google.com/docs/firestore/quotas)
-  }
-
-  hasUnflushedEvent() {
-    return !!this.buffer.length;
-  }
-
-  async flush() {
-    if (!this._flushing && (this.buffer.length >= 3 || (new Date()) - this._previousFlushTime >= 5 * 60 * 1000)) {
-      this._flushing = true;
-      await this._flush();
-      this._flushing = false;
-    }
-  }
-
-  async _flush() {
-    if (!this.db.backendController._initialized) {
-      return;
-    }
-
-    try {
-      await this.addDocs(this.buffer);
-      this.buffer = [];
-      await this.saveBuffer();
-    } 
-    catch (err) {
-      throw new Error(`Failed when flushing: ${err.message}`);
-    }
-
-    this._previousFlushTime = new Date();
   }
 
   addEvent = (name, data) => {
@@ -89,9 +48,5 @@ export default class UserEventContainer extends FirestoreContainer {
         logError(err);
       }
     })();
-  }
-
-  async saveBuffer() {
-    return this.db.backendController.practiceManager.externals.storage.set(this._mementoKeyName, this.buffer);
   }
 }
